requirements_test.txt
-r requirements.txt
factory-boy==1.3.0
mock==1.0.1
six==1.3.0
wsgiref==0.1.2
django-nose
